[native] Prevent tooltip from appearing on spoiler tap if spoiler is not yet revealed
Summary:
This diff handles refactoring the MarkdownSpoiler component to be able to detect when a user taps to reveal. This is necessary in order to do two things: 1) If the spoiler has not been revealed, and a user taps on it, we should prevent the tooltip from appearing and just display the message contents, and 2) If the spoiler is already revealed, we want to allow the tooltip to appear on tap.

ENG-2072: https://linear.app/comm/issue/ENG-2072/maintain-the-spoiler-state-upon-opening-textmessagetooltipmodal

Depends on D5515

Test Plan:
Create a chat with two separate text messages, each with their own spoilers. The main test is to check whether the tooltip appears when a revealed spoiler / non-spoiler part of the message is tapped, and whether it appears when a spoiler that has not yet been revealed yet is tapped.

{F221961}

Reviewers: atul, ginsu

Subscribers: ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5539