[services] Tunnelbroker - Notifications config initialization
Summary:
In this diff, we initialize the static configuration as a struct for the internal notifications library to avoid passing repeated parameters on every notification-sending call. We are using a lazy static initialization and parameters from the Tunnelbroker's configuration file.
This diff is related to the abandoned D5329. The static configuration is used in the D5642 diff updates.

Linear task: [[ https://linear.app/comm/issue/ENG-1737/change-to-use-builder-approach-and-thread-local-in-rust-notification | ENG-1737 ]]

Test Plan:
1. Service is successfully built.
2. The diff D5641 where the function is used the service is successfully built.

Reviewers: jon, tomek, varun

Reviewed By: jon, tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5641