[lib] add legal policies list
Summary:
Add hardcoded list of policies that users will need to acknowledge.

If any new policy will be introduced only operation needed will be updating this object.

I decided to use `object` instead of the only array with string values, maybe in future there will be a need to distinguish between them, make some operations only on one and it will be easier to access (eg. now on running a script and choosing which policy requires re-acknowledge it can be based on object prop).

Test Plan: `yarn flow check`

Reviewers: atul, marcin, ashoat, tomek

Reviewed By: atul, marcin, ashoat, tomek

Subscribers: ashoat, atul, abosh

Differential Revision: https://phab.comm.dev/D5306