[keyserver] update acknowledgment table in login was performed from proper source
Summary:
Code that updates this table when a login is received with the correct `source` set.

Note about older clients which don't have code for passing source: I think we should allow them to login but not make any changes in this table, they will be able to use the app but we will have only fully truthful data about their acknowledgments.

Test Plan: Check if table was updated for logins from the native form and was not changed for any other login (eg. web, restoring outdated cookie).

Reviewers: atul, marcin, ashoat

Reviewed By: ashoat

Subscribers: ashoat, atul, abosh

Differential Revision: https://phab.comm.dev/D5312