[keyserver] add migration to populate `policy_acknowledgments` table
Summary:
A solution suggested here: D5309#158058. Creates an entry for every already existing user.

Note: there are two users (ashoat and commbot) who will be present in the database on fresh creation and they will not have an entry in the `acknowledgment_table` but I am not sure if this is needed.

Test Plan:
1. Run migration
2. this code should add rows to `policy_acknowledgments` for all users with terms of use policy, current time, and confirmed set to 1.

Reviewers: tomek, marcin, atul

Reviewed By: tomek, atul

Subscribers: ginsu, ashoat, atul, abosh

Differential Revision: https://phab.comm.dev/D5663