[keyserver] added target_message to messages table
Summary:
added `target_message` to messages table. `target_message` is the id of the message that is being reacted to. Also added the index to `target_message` column and this is to optimize how we will be fetching reaction messages for a certain message

---

Linear Task: [[ https://linear.app/comm/issue/ENG-2265/modify-messages-table-to-include-targetmessageid-column | ENG-2265 ]]

Test Plan:
Please see the screenshot to see that `target_message` is added to the `messages` table

{F248488}

Also dropped and created my local DB to test the setup works as expected. Tested the stack locally in my environment to make sure everything still ran as expected/nothing crashed.

Reviewers: atul, tomek, rohan, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5658