[web] Typeahead overlay visual
Summary:
Component that will be displayed when user tries to mention other user (by typing @ in chatbar).
Currently, it is not created anywhere, but code that does it will be added in next diffs.
Component takes list of actions and it's absolute position that will be calculated outside of it.
Notice it isn't positioned correctly yet, as it's work in progress.

Relevant linear tasks:
https://linear.app/comm/issue/ENG-2096/typeahead-in-tooltip-for-mentions
https://linear.app/comm/issue/DES-1/design-typeahead-for-mentions-on-web#comment-f40bb8c5

Test Plan:
I already have most of the outer logic completed, so I was able to test if it's rendered properly.
Video below:
{F249233}

Reviewers: tomek, kamil, atul

Reviewed By: tomek, atul

Subscribers: ashoat, abosh, atul, kamil, tomek

Differential Revision: https://phab.comm.dev/D5660