[lib] Update `rawMessageInfoFromClientDB` preconditions to allow `MULTIMEDIA` messages.
Summary:
`invariant` at start of `rawMessageInfoFromClientDB` currently always fails for messages of type `MULTIMEDIA`. This diff updates the invariant to accept either `MULTIMEDIA` or `IMAGES` messages.

Actual handling of `MULTIMEDIA` messages will come in a subsequent diff... this is mostly just to put up changes that can be landed as-is to reduce number of changes I have to juggle locally.

Test Plan: Things continue to work as expected.

Reviewers: tomek, marcin, ginsu, rohan, kamil, O2 Blocking Reviewers, ashoat

Reviewed By: O2 Blocking Reviewers, ashoat

Subscribers: ashoat, abosh

Differential Revision: https://phab.comm.dev/D5636