[desktop] Double-click to maximize or minimize
Summary: After hiding the default title bar (because it doesn't look good), we lose the functionality of double clicking. There's no proper way to do it with electron, so I detect the double click on the top bar and manually activate the correct action. Clickable elements on the top bar have to stop propagation of the double click so users don't maximize by mistake.

Test Plan:
- Test if double click to (un)maximize works
- Change mac settings and check if minimize works
- Change mac settings and check if "Do nothing" does nothing

Reviewers: tomek, atul, abosh, kamil

Reviewed By: tomek, atul

Subscribers: kamil, ashoat

Differential Revision: https://phab.comm.dev/D5631