[lib] Refactor getThreadOtherUsers and all its occurrences
Summary:
I refactored `getThreadOtherUsers`: changed its name so it matches methods around and
make it return (Relative)MemberInfo's instead of IDs. It was only used in `thread-utils.js` and
it was easy to changed places where it was called. I need it in that form in typeahead, so I figured
it's fine to extend that function instead of creating another almost identical one.

Test Plan:
Tested in furhter diffs.
Visible and works correctly returning all users excluding ashoat and viewer.
Performed sanity check if I didn't break the function (traced one of the functions (getSingleOtherUser) that called it and it performed as expected (in Relationship Prompt))

Reviewers: tomek, kamil, atul

Reviewed By: tomek

Subscribers: ashoat, atul, kamil, tomek

Differential Revision: https://phab.comm.dev/D5673