Allow for tooltip to be hidden when revealing spoiler while maintaining styles
Summary:
Currently, in D5539, we introduce `GestureTouchableOpacity` in
order to block the tooltip from appearing when we want to reveal the
spoiler. There are several problems with this approach:

1) Pressing a spoiler once it is already revealed still prevents the
tooltip from appearing.

2) More importantly, the styling is completely different to what is within
the text message. The text size, color, and also the y-axis of the text is
all unexpected (Context:
https://linear.app/comm/issue/ENG-2272/gesturetouchableopacity-changing-styles).

Based on some experimenting, this allows us to have the two desired
behaviors - the spoiler is revealed without showing the tooltip, and the
styling once revealed is consistent.

I plan to resolve the need for the eslint ignore comment before landing,
but I'm putting this diff up to allow for reviewers to repro locally to
test, and also to figure out the reason for these changes fixing things.
Speaking with @atul briefly, we think it may be some dependency handling
fault from our end.

Test Plan:
Demo video is attatched below. The test plan:

1) Restart `keyserver` and `native` with `yarn dev`
2) Simulate a spoiler reveal action by tapping on a spoiler
3) Simulate a general tap on a message outside of spoilers to expect
tooltip
4) Simulate a tap on a spoiler (after revealed) to expect tooltip
5) Simulate a long-press on a message to expect the tooltip

Reviewers: atul!, ginsu

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5688