[native] introduced react option to the message tooltip
Summary: introduced the `react` option to the text message and the multimedia message tooltips. I decided not to add it to the robotext message because I feel like users should not be able to react to those types of message, but if we want to change that it will be a very easy fix.

Test Plan: Please watch the demo below to see how the `react` option looks like

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5783