Move ThreadDraftUpdater component to lib
Summary: This differential moves ThreadDraftUpdater component to lib, so that it can be used on web app as well.

Test Plan: Draft moving functionality should remain on native. Open web app and native app on the same clien. Start typing message on the sidebar on native, then create the same sidebar on web. Draft should still be visible on native.

Reviewers: atul, tomek, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D5707