Move Store Loaded to redux
Summary: This differential introduces store loaded to redux and makes sure it is not persisted. There are components that check `storeLoaded` flag to ensure data from SQLite was inserted to redux. Setting this flag in react state after redux actions are dispatched introduces some sort of race condition. Therefore we should keep this flag in redux and set it in reducer in response to the same action that inserts SQLite data.

Test Plan: Make sure eslint and flow do not complain. add code in `reducer` in native that sets this flag to true. Add logging to make sure this flag is `false` after re-starting application.

Reviewers: tomek, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5709