[native] Fix getStateFromNavigatorRoute calls
Summary:
Linear issue: https://linear.app/comm/issue/ENG-2316/threadscreenpruner-problem
This code was trying to obtain route for `Chat` but was obtaining route for `Profile`. This way we can be sure that we are actually finding what we are looking for.

Test Plan: Run ios simulator, see that both on login screen and when user is logged in no errors arise. I also logged the value that is being returend to be sure this component is being rendered at all.

Reviewers: atul, tomek, kamil, bartek, ginsu, przemek, ashoat

Reviewed By: ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5705