[web] Creating typeahead overlay in ChatInputbar
Summary:
Final diff for MVP typeahead for web.
We lack some nice UX from design, but it can be done in future diffs after
we implement typeahead on native.
Nice to haves (that we don't have yet):

- keyboard support (choosing with arrows, confirming with Enter, cancelling with Esc), for now you gotta click username with your mouse
- @ing people that are not in a channel thread would create a modal asking if you want to invaite them

See a video for how it feels and works:
{F260093}

Test Plan:
See a video above.
I played with it a lot during development and it feels alright, but would appreciete any feedback after it is landed and you can test it yourself.

Reviewers: tomek, atul, kamil

Reviewed By: tomek

Subscribers: ashoat, kamil, atul, tomek

Differential Revision: https://phab.comm.dev/D5722