[services][blob] Add AWS SDK config
Summary:
Added a function to retrieve SDK config, required for accessing
the S3 and DynamoDB SDKs. It is unused yet, but will be needed
later.

Conditionally, if `COMM_SERVICES_SANDBOX` environment variable is set,
it uses localstack, which reflects the
[current C++ implementation](https://github.com/CommE2E/comm/blob/2007dbe05b7d41d1603b74ccbef515e6942b1d21/services/lib/src/DynamoDBTools.cpp).

Depends on D5677

Test Plan: This function is tested a few diffs later, when S3 / DynamoDB are used.

Reviewers: varun, jon, tomek, marcin, ashoat

Reviewed By: varun, tomek, ashoat

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D5678