[services][blob] Add methods to manage the blob table
Summary:
Implemented methods for managing the `blob-service-blob` table. They will be needed in subsequent diffs.

This is mostly the same approach as in Identity Service, except I use `anyhow` to handle errors for now, for the sake of simplicity. Proper error handling is going to be introduced later.

[Link to the C++ counterpart](https://github.com/CommE2E/comm/blob/32d6628a1db51c4f1432fa471bf6508cd331f7d7/services/blob/src/DatabaseManager.cpp)

Part of [[ https://linear.app/comm/issue/ENG-2300/access-s3-and-dynamodb-in-blob-service | ENG-2300 ]]

Test Plan: Called these locally and ensured that entities are properly inserted/read/removed from database.

Reviewers: varun, jon, tomek, marcin

Reviewed By: varun, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D5693