[services] Tunnelbroker - Remove checkpoints from the database
Summary:
In past research, we decided to switch from the time-based message checkpoints to the unique message Ids because of the complexity of the time synchronization between the service instances. The checkpoints-related fields are still in the database model and should be removed.

The full context is in [[ https://linear.app/comm/issue/ENG-1158/message-delivery-confirmation-messageids-batching-in-tunnelbroker | ENG-1158 ]].

Linear task: [[ https://linear.app/comm/issue/ENG-2357/remove-time-based-messages-checkpoints-from-the-database-model | ENG-2357 ]]

Test Plan: Service is successfully built.

Reviewers: jon, varun, tomek

Reviewed By: tomek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5733