[Nix Docs] Remind user to restart terminal
Summary:
Make it clear in documentation that the terminal should be restarted
or a new terminal should be opened to apply nix changes. This is needed so that
the nix changes can be sourced and applied to the new session.

The install script from nix also emits this advice. However, that output gets pushed
up severely by all of the output we emit when adding defaults to `/etc/nix.conf`

Although it's possible to source a file, the file location seems to be
dependent on how nix was installed.

https://linear.app/comm/issue/ENG-2358

Test Plan: N/A

Reviewers: ashoat, varun

Reviewed By: ashoat

Subscribers: tomek, atul

Differential Revision: https://phab.comm.dev/D5741