[native] Remove gRPC related C++ code and packages from iOS and Android
Summary:
This diff introduces removing the C++ gRPC-related code and packages due to the switching to use Rust gRPC client in a native.
The following files, includes and links were removed as deprecated or to not leave the empty files as the implementation would be completely changed:
- GlobalNetworkSingleton.cpp - The file will be completely changed, we don't want to store an empty file;
- NetworkModule.cpp - The file will be completely changed, we don't want to store an empty file;
- SocketStatus.h - Deprecated as a part of the C++ gRPC;
- Client.cpp - Deprecated as a part of the C++ gRPC;

The followings packages were removed:
- gRPC-C++;
- Pod-patch - We used it to patch the gRPC-C++ package, we don't use C++ gRPC anymore;

Linear task: [[ https://linear.app/comm/issue/ENG-2276/removing-grpc-c-library-openssl-on-android-pod-patch-on-ios | ENG-2276 ]]

Test Plan: Native iOS and Android apps are successfully built.

Reviewers: atul, tomek, jon, ashoat

Reviewed By: atul, tomek, ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5706