[native] 2/N Disable 'Delete' button on delete acconut modal when password input is empty
Summary:
This diff covers the native side of disabling the "Delete account" button when a user has yet to input anything.
Curiously, the disabled feature on native appears to not gray-out the button but only not allow users to click it (see test video)
[[ https://linear.app/comm/issue/ENG-2014/unknown-error-when-trying-to-delete-account-change-password-with-a  | ENG-2014 ]]

Test Plan:
Verify that the button is disabled (unclickable) when text input is empty.

https://www.loom.com/share/b89ddf1249eb4b05aa56fee974a1552b

Reviewers: atul, rohan, ginsu

Reviewed By: atul, rohan

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5778