[native] Maintain the revealed state of spoilers
Summary:
This diff handles storing the spoiler reveal state so when a user taps on a message once the spoiler is revealed, it won't "re-hide" (the behavior that is present further down in this stack). With this diff, we can now keep track of what spoilers are revealed and what spoilers are still hidden, and the style of the `Text` component changes once a user reveals the spoiler.

ENG-2072: https://linear.app/comm/issue/ENG-2072/maintain-the-spoiler-state-upon-opening-textmessagetooltipmodal

Depends on D5515

Test Plan:
Create a chat with two separate text messages, each with their own spoilers. The main test is to check whether the spoiler "reveal" state is maintained when the tooltip appears.

{F221961}

Reviewers: atul, ginsu, ashoat

Reviewed By: atul, ashoat

Subscribers: michal, ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5539