Revert "[native] Remove `sortBy` property from `fetchPhotos`"
Summary:
This reverts commit 091b31930a107ddbbb73be29ccdf8649235e2e62.

Context: https://linear.app/comm/issue/ENG-1481/revisit-medialibrary-sortby
Additional context: https://linear.app/comm/issue/ENG-2111/max-sees-old-images-ahead-of-recent-ones-in-image-gallery

Basically @max and I and others noticed that the sort order of images in the `MediaGalleryKeyboard` appears incorrect. This seems to be the case both on Android and iOS (see Linear issues for further discussion).

The changes in D4836 were intended to address an issue where media was being "pushed to the front" when viewed/edited because `MediaGalleryKeyboard` was sorting by `modificationTime`.

However, after that diff was landed I noticed that the order of media in `MediaGalleryKeyboard` was not as expected on my personal iOS device. In addition, after pushing build 159 to prod on Android, I got feedback that media appeared to be sorted in the wrong order... specifically that it seemed like media was being sorted alphabetically based on the names of the folders in which they lived.

This diff reverts the sort order to what it was in build 131 (last public App Store/Play Store release)

Test Plan:
Pushed release build to my phone (iOS) and things appear as expected.

But generally this should match behavior of existing "public" release, so I think this is a safe/predictable change.

Reviewers: ashoat, tomek, marcin

Reviewed By: ashoat

Subscribers: max

Differential Revision: https://phab.comm.dev/D5781