[lib] introduced unsettingReaction field to reaction type
Summary: introduced `unsettingReaction` field to reaction type. `unsettingReaction` is a `boolean` that describes wheter the reaction in the `reaction` field should be set or unset

Test Plan: `flow`, and nothing crashes locally on native/web.  also tested adding this field with the exisiting message liking infrastructure in my local stack and was able to get it to appear in the DB.

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5801