[services] Tunnelbroker - Removing rudimental `messageID` from messages from client to Tunnelbroker
Summary:
This diff removes the rudimental `messageID` field in the Tunnelbroker protocol buffers file for the Client to Tunnelbroker messages communication.

When the client sends messages to the Tunnelbroker the Tunnelbroker sends back message IDs to the client to confirm they are processed. Message IDs are generated on the Tunnelbroker side and are unique IDs for each message. We should remove the messageID from the protobuf file for the messages that come from the client to the Tunnelbroker as it's not used. That field came from the reason that the structure of the messages should be the same for the Tunnelbroker -> Client and Client -> Tunnelbroker communication, but it's changed due to the API changes, and this rudimental field is left and should be removed. The API description in an [[ https://www.notion.so/commapp/Tunnelbroker-028bdd578e4c487fb570dbb862952ec6#ecbe3fcc7a9345298318d269caf362ed | architecture review ]].

Linear task: [[ https://linear.app/comm/issue/ENG-2405/removing-rudimentals-from-the-protobuf-file | ENG-2405 ]]

Test Plan:
1. Tunnelbroker service is successfully built.
2. Tests for the Tunnelbroker are passed.
3. Manual testing of the API is passed.

Reviewers: tomek, varun, jon, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5808