[services] Tunnelbroker - Removing rudimental `ping` message from the protobuf
Summary:
This diff removes the rudimental `ping` message from the protobuf.

The "pinging" message mechanism was changed to the [[ https://linear.app/comm/issue/ENG-1842/consider-using-of-grpc-keep-alive-instead-of-ping-messages | gRPC keep-alive mechanism ]] and the server implementation was changed in D5713. The empty "ping" message should be removed from the protobuf API as it's not used anymore.

Linear task: [[ https://linear.app/comm/issue/ENG-2405/removing-rudimentals-from-the-protobuf-file | ENG-2405 ]]

Test Plan:
1. Services are successfully built.
2. CI tests passed.

Reviewers: jon, tomek, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5809