[lib] Prevent spoiler text from appearing in MessagePreview
Summary:
With the implementation of Spoilers on both `web` and `native`, we want to prevent the spoiler text from being revealed in `MessagePreview` on both platforms. This diff handles replacing the text enclosed with spoiler characters with three black squares, as discussed in the parent tasks for Spoilers, [ENG-1472](https://linear.app/comm/issue/ENG-1472/spoiler-quotes-in-markdown). We previously thought about replacing the spoilers with black square emojis.

The idea with this diff is to "intercept" the raw message text before it is parsed for markdown rules, so we can replace the text enclosed within `|| ... ||` before the spoiler characters are removed for the message preview.

[Context Diff](https://phab.comm.dev/D693)

Addresses [ENG-2193](https://linear.app/comm/issue/ENG-2193/prevent-spoiler-text-from-appearing-in-messagepreview-[web])
Addresses [ENG-2055](https://linear.app/comm/issue/ENG-2055/prevent-spoiler-text-from-appearing-in-messagepreview-[native])

Test Plan:
[Web]:
1. Restart `keyserver` and `web`
2. Observe how spoilers are displayed in the message preview before the change
3. Observe after the change
4. Ensure other markdown functions as before
{F259297}

{F259296}

[Native]:
1. Restart `keyserver` and `native`
2. Observe how spoilers are displayed in the message preview before the change
3. Observe after the change
4. Ensure other markdown functions as before
5. Check how the change looks in both light and dark mode
{F259301}

{F259300}

{F259298}

{F259299}

Reviewers: atul, ginsu, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D5715