[web] Add unread indicator within subchannel model
Summary:
This diff adds styling so that unread subchannel messages are consistent with the rest of the web app, that is, they are made 'semi-bold' and that the color of the subchannel name, sender, text preview, and time are made white.

Deciding to change the color of all the subchannel elements was something I thought was appropriate but I'm happy to tinker some more with only color-changing certain elements.

This is one half of the linear task [[ https://linear.app/comm/issue/ENG-2089/add-unread-indicators-to-sidebar-modal-on-web | ENG-2089 ]]. The next step is to do this process for unread indicators for the sidebar modals.

Test Plan:
| Unread SubChannel  | Read Subchannel
| -----  | -----
| {F272286}  | {F272289}
| {F272288} | {F272290}

Reviewers: atul, ginsu, rohan

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D5810