[native] [27/40] RN 0.70: Patch react-native-camera to fix ViewPropTypes crash
Summary:
Our old version of `react-native-camera` was erroring due to an import of `ViewPropTypes` from `react-native`, which has been deprecated. I tried to update it, but that didn't fix the issue, so I added a patch. Then I realized the update broke the Android build, so I took it out.

This package appears to have been deprecated sometime in the last year. [ENG-1520](https://linear.app/comm/issue/ENG-1520/explore-using-react-native-vision-camera-for-video-capture) tracks the long-term solution, but I don't have the cycles to handle that right now unfortunately.

Depends on D5921

Test Plan:
1. Confirm JS error due to `ViewPropTypes` import no longer occurs
2. Tested along with whole stack: [test plan](https://www.notion.so/commapp/Test-plan-for-React-Native-0-70-upgrade-718fa7e0fe4348099b4c169eb121da16) (includes testing camera functionality)

Reviewers: tomek, bartek

Subscribers: atul

Differential Revision: https://phab.comm.dev/D5922