[lib] introduced reactions field to chatMessageItems
Summary: introduced `reactions` to `chatMessageItems` in the chat selectors. For this introduction, I created a map with the `targetMessageID` as the key and an array of reaction message infos as the values.

Test Plan:
Will be doing more tests in subsequent diffs. Ran mobile/web app locally and nothing crashes. I also was able to get to the point where I could send and reaction and it would render on my local
stack (see demo video below)

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5811