[web] Standardize spoiler color
Summary: Two things here - the first is that we probably want to make the color of spoilers the same as it is on native, `#33332C`. The second thing is the current spoiler color on received messages is impossible to differentiate from the message bubble.

Test Plan:
Will link before/after on `web`

| **Before** | **After**    |
| {F275071} | {F275072} |
| {F275104} | {F275105} |

Reviewers: atul, tomek, ginsu

Reviewed By: atul, ginsu

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5836