[lib] modified messagePreviewText to handle reaction message types
Summary: Made changes to `messagePreviewText` to handle reaction message types. We will need this for the `messageTitle` function in the reaction message spec. For now, I am going with a generic text format of "someone reacted to a message", but will probably change the content when we get further along with the message liking implementation.

Test Plan: Ran web and mobile app locally and nothing crashes/will be testing more in subsequent diffs

Reviewers: atul, rohan

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D5689